MSVC, Reflection Fixed MSVC support (idk why asan is triggering in some places in MSVC that it doesn't in G++)
Minor refactor to Reflection

+++ src/impl/render/Renderer.cpp
@@ -450,7 +450,7 @@ void Renderer::DrawScene(
 			renderData.Size,
 			renderData.Transform,
 			renderData.FaceCulling,
-			drawInfos.size()
+			static_cast<int32_t>(drawInfos.size())
 		);
 	}
 }
